Automate Georgia
* completely reworked code in crawl to use bs4 and common.cache_request with POST (instead of selenium); moved into georgia main.py
* note that the random 1-3 second wait between post requests was retained
* moved parse code into georgia main.py since it's now a manageable amount of code to keep in one file
* renamed parse_html to parse_county, removed file load section, and tweaked email portion to use common.decode_email
* no changes to parse_addr_line, parse_contact_line, parse_contact
* added lines to main to sort by locale and use common.diff_and_save
* diff_and_save showed 1 email change and 1 official name change when run against old data sorted by locale
* updated Analysis.ipynb
